pyproject.toml
[tool.poetry]
name = "kryten-playlist"
version = "0.3.2"
description = "Kryten playlist management service - handles video queue, playlist operations, and catalog enrichment with LLM"
authors = ["Kryten Robot Team"]
readme = "README.md"
license = "MIT"
homepage = "https://github.com/grobertson/kryten-playlist"
repository = "https://github.com/grobertson/kryten-playlist"
documentation = "https://github.com/grobertson/kryten-playlist/blob/main/README.md"
keywords = ["cytube", "moderation", "chat", "nats", "microservices", "catalog", "llm", "enrichment"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Communications :: Chat",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Multimedia :: Video",
]
packages = [{include = "kryten_playlist"}]
include = ["systemd/*.service", "systemd/README.md", "INSTALL.md"]
exclude = ["config.json", "config-*.json", "logs/*", "*.log"]

[tool.poetry.dependencies]
python = "^3.10"
kryten-py = ">=0.9.13"
fastapi = "^0.115.0"
uvicorn = "^0.30.0"
aiosqlite = "^0.20.0"
jinja2 = "^3.1.4"
httpx = "^0.27.0"
click = "^8.1.0"
json_repair = "^0.54.3"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
pytest-asyncio = "^0.24.0"
pytest-mock = "^3.14.0"
pytest-httpx = "^0.30.0"
black = "^23.0.0"
ruff = "^0.1.0"
mypy = "^1.7.0"

[tool.poetry.scripts]
kryten-playlist = "kryten_playlist.__main__:main"
kryten-catalog = "kryten_playlist.catalog.cli:cli"
kryten-ingest = "kryten_playlist.catalog.cli:ingest_standalone"
kryten-enrich = "kryten_playlist.catalog.cli:enrich_batch_standalone"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
addopts = "-v"
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]

[tool.black]
line-length = 100
target-version = ['py310']

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N"]
ignore = ["E501"]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["E402", "F401"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
